+++ backend/jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/formengine/entity/FormFieldMeta.java
@@ -61,6 +61,9 @@ public class FormFieldMeta {
     @ApiModelProperty("default value")
     private String defaultValue;
 
+    @ApiModelProperty("searchable")
+    private Integer searchable;
+
     @ApiModelProperty("status")
     private Integer status;